Implement DamageTracker and integrate it into MatchStatTracker
DamageTracker is responsible for tracking damage and includes a method for getting who should be credited in an assist.

+++ Assets/Assemblies/MatchLogic/DamageContribution.cs
@@ -4,6 +4,6 @@ namespace Resonance.Assemblies.Match
     {
         public ulong attackerId;
         public float damageAmount;
-        public float timestampUnixTimeMs;
+        public long timestampUnixTimeMs;
     }
 }
